requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
phidata>=2.4.25
openai>=1.3.7
duckduckgo-search>=3.9.6
yfinance>=0.2.22
requests>=2.31.0
